requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.1.0
jinja2==3.1.6
